Fix lint errors in Calendar.jsx and useCalendarEvents.js
